pyproject.toml
[project]
name = "bliq"
version = "0.1.0"
description = "Dataset catalog with metadata storage and efficient querying"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.1",
    "httpx>=0.27.0",
    "requests>=2.31.0",
    "uvicorn>=0.38.0",
    "azure-identity>=1.15.0",
    "azure-storage-blob>=12.19.0",
    "boto3>=1.34.0",
    "duckdb>=0.10.0",
    "pandas>=2.0.0",
    "pyarrow>=15.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "peewee>=3.18.2",
]

[project.optional-dependencies]
postgresql = [
    "psycopg2-binary>=2.9.0",
]

[project.scripts]
bliq = "bliq.cli:cli"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ipython>=9.2.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "pytest>=8.4.1",
    "ruff>=0.13.1",
]


[tool.uv]
default-groups = ["dev"]
